pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dashboard-generator"
version = "1.0.0"
description = "AI Data Analyzer with Streamlit"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "streamlit==1.31.1",
    "plotly==5.19.0",
    "pandas==2.2.1",
    "requests==2.31.0",
    "groq==0.5.0",
    "python-dotenv==1.0.0",
    "openpyxl==3.1.2",
    "xlrd==2.0.1",
    "pdfplumber==0.11.0",
    "numpy==1.26.3"
]

[tool.setuptools]
packages = []